Implement get_q_update and qLearnMinibatch (ish)
Number of things:

Switched to the new retro package from openAI. its cool but you have to
get the roms yourself, AND ones that it will approve of (i cheated and
edited the SHA1 file in the package to get my SpaceInvaders to work).

get_q_update and qLearnMinibatch is 'working'. I say 'ish' because I am
currently defaulting to an Adam optimiser which may not be the correct
option. However I am still using AEGreedy and compiling it everytime,
so we'll see as more features get added. In the meantime the asserts
are just commented out as i try this.