Switch to pyTorch
Should have done this at the start. Keras was not giving me the
flexability i need to get this (hopefully) working. AND the original DQN
was made in torch7. The gradient and weight manipulation is also
included (though i am not sure I have the backprop set correctly)

With the move its actually a little bit faster, but im still not seeing
the performance i was hoping from the agent. I'm hoping that with more
steps ill see the reward improving.